% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathsReport.R
\name{cadia.paths}
\alias{cadia.paths}
\title{Generates a summary of the pathways used in CADIA}
\usage{
cadia.paths()
}
\value{
The list of pathways used in CADIA and their KEGG IDs
}
\description{
Generates a summary of the pathways used in CADIA
}
