% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{causalDisturbance}
\alias{causalDisturbance}
\title{Calculates enrichment analysis of KEGG pathways based on causal disturbance
model. For more information, see publication doi>10.1145/3107411.3107488}
\usage{
causalDisturbance(deIDs, allIDs, iter = 2000, alpha)
}
\arguments{
\item{deIDs}{a list of differentially expressed genes}

\item{allIDs}{list of all genes in the experiment}

\item{iter}{Iteration round for sampling causal disturbance}
}
\value{
list containing pathway names, number of pathway nodes and edges,
        Over-representation P-values, size of differentially expressed genes
        in the pathway, and causal disturbance
}
\description{
Calculates enrichment analysis of KEGG pathways based on causal disturbance
model. For more information, see publication doi>10.1145/3107411.3107488
}
