% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDAG.R
\name{DAGprocessor}
\alias{DAGprocessor}
\title{Finds an underlying DAG of a given graph pathway
Uses a greedy hueristic}
\usage{
DAGprocessor(inputGraphDAG)
}
\arguments{
\item{inputGraphDAG}{the graph of the pathway to be processed}
}
\value{
processed graph without anyloops or self loops
}
\description{
Finds an underlying DAG of a given graph pathway
Uses a greedy hueristic
}
