% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExtractorKEGG.R
\name{geneKEGGListExtractor}
\alias{geneKEGGListExtractor}
\title{Extracts all the genes in a given list of pathways}
\usage{
geneKEGGListExtractor(listOfPathwayFiles)
}
\arguments{
\item{listOfPathwayFiles}{An input containing a list of pathways where the
original files are in the format of KGML.}
}
\value{
A list of genes using associated KEGG ID
}
\description{
Extracts all the genes in a given list of pathways
}
