% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneReport.R
\name{geneReport}
\alias{geneReport}
\title{Returns the list of DE genes in a set of pathways}
\usage{
geneReport(deIDs, pathwayList = NULL)
}
\arguments{
\item{deIDs}{a list of differentially expressed genes}

\item{pathwayIDs}{a list KEGG IDs of pathways for which we want the DEGs}
}
\description{
Returns the list of DE genes in a set of pathways
}
