% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newpathCentrality.R
\name{newpath.centrality}
\alias{newpath.centrality}
\title{Calculates the Source/Sink Centrality of a pathway}
\usage{
newpath.centrality(adj.matrix, alpha, beta)
}
\arguments{
\item{adj.matrix}{the adjacency matrix of a graph to be analyzed}

\item{alpha}{is the decay factor, similar to the one in Katz centrality}

\item{beta}{is the relative effect of the sinks vs sources.}
}
\value{
A matrix in which the row i hold the vector of influence of the
node i on all the other nodes.
}
\description{
Calculates the Source/Sink Centrality of a pathway
}
