% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathCountKatz.R
\name{pathCounter.katz}
\alias{pathCounter.katz}
\title{Find the total number of paths in a DAG}
\usage{
pathCounter.katz(adjMatDAG, eye, alpha)
}
\arguments{
\item{adjMatDAG}{An adjacency matrix of a DAG}

\item{eye}{is the identity function}
}
\value{
the total number of paths in the graph
}
\description{
Takes an adjacency matrix input
}
