% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbanceCalculatorKatz.R
\name{pathSampler.katz}
\alias{pathSampler.katz}
\title{Calculates the distrution of pathway disturbance}
\usage{
pathSampler.katz(pathwayRefGraphExpanded, iterationNo, deKID, allKID)
}
\arguments{
\item{pathwayRefGraphExpanded}{An KEGGGraph object of a pathway}

\item{deKID}{KEGG ID of the differentially expressed genes}

\item{allKID}{reference of all the genes with their respective KEGG IDs}

\item{interationNo}{number of rounds of sampling}
}
\value{
a  list of two objects. First is the sampling data for
causal disturbance. The second one is the actual causal disturbance
}
\description{
Calculates the distrution of pathway disturbance
}
