% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbanceCalculatorNewPath2.R
\name{pathSampler.newPath2}
\alias{pathSampler.newPath2}
\title{Calculates the distrution of pathway disturbance}
\usage{
pathSampler.newPath2(pathwayGraph, iterationNo, deKID, allKID, alpha)
}
\arguments{
\item{deKID}{KEGG ID of the differentially expressed genes}

\item{allKID}{reference of all the genes with their respective KEGG IDs}

\item{pathwayGraph2}{An KEGGGraph object of a pathway}

\item{interationNo}{number of rounds of sampling}
}
\value{
a  list of two objects. First is the sampling data for
causal disturbance. The second one is the actual causal disturbance
}
\description{
Calculates the distrution of pathway disturbance
}
