% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathSampler.R
\name{pathSampler}
\alias{pathSampler}
\title{Calculates the distribution of pathway disturbance}
\usage{
pathSampler(inputGraph, deKID, iterationNo, alpha, beta, statEval = 1)
}
\arguments{
\item{inputGraph}{A graphNEL object of a pathway}

\item{deKID}{The name of the nodes in the subset of interest. e.g. DEG}

\item{alpha}{the dampening factor for Source/Sink Centrality}

\item{beta}{the relative Source vs Sink factor}

\item{statEval}{Choose 1 for product-based, 0 for summation-based}

\item{interationNo}{number of rounds of sampling}

\item{iter}{the number of iterations for causal disturbance}
}
\value{
the probability of observing a subset of
}
\description{
Calculates the distribution of pathway disturbance
}
