% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayDist3.R
\name{processPathway}
\alias{processPathway}
\title{Calculates some information regarding the pathways}
\usage{
processPathway(pGraph, Name, deKIDs, allKIDs, keggRef, iter)
}
\arguments{
\item{pGraph}{a GraphNEL object containing the pathways}

\item{Name}{the pathway name}

\item{deKIDs}{list of differentially expressed genes}

\item{allKIDs}{list of all genes with their associated KEGG IDs}

\item{keggRef}{list of reference genes from KEGG}

\item{iter}{the number of iterations for causal disturbance}

\item{pathID}{KEGG ID of an input pathway}
}
\description{
Calculates some information regarding the pathways
}
