% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{causalDisturbance}
\alias{causalDisturbance}
\title{Calculates enrichment analysis of KEGG pathways based on causal disturbance
model. For more information, see publication doi>10.1145/3107411.3107488}
\usage{
causalDisturbance(deIDs, allIDs, iter = 2000, alpha = 0.1, beta = 1,
  statEval = 1, fdrMethod = "BH", verbose = F)
}
\arguments{
\item{deIDs}{a list of differentially expressed genes}

\item{allIDs}{list of all genes in the experiment}

\item{iter}{Iteration round for sampling causal disturbance}

\item{alpha}{Decay factor for Source/Sink centrality, default 0.1.}

\item{beta}{Relative importance of sink versus source}

\item{statEval}{0 for summation of source/sink centrality 1 for product}

\item{fdrMethod}{choice of false disovery rate method. inherited from p.adjust}

\item{verbose}{printing pathway processing report}
}
\value{
list containing pathway names, number of pathway nodes and edges,
        Over-representation P-values, size of differentially expressed genes
        in the pathway, and causal disturbance
}
\description{
Calculates enrichment analysis of KEGG pathways based on causal disturbance
model. For more information, see publication doi>10.1145/3107411.3107488
}
