% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathCount2.R
\name{pathCounter}
\alias{pathCounter}
\title{Find the total number of paths in a DAG}
\usage{
pathCounter(adjMatDAG, eye)
}
\arguments{
\item{adjMatDAG}{An adjacency matrix of a DAG}

\item{eye}{is the identity function}
}
\value{
the total number of paths in the graph
}
\description{
Takes an adjacency matrix input
}
