% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayEnrichment.R
\name{processPathway}
\alias{processPathway}
\title{Calculates some information regarding the pathways}
\usage{
processPathway(pGraph, Name, deIDs, allIDs, iter, alpha, beta, statEval)
}
\arguments{
\item{pGraph}{a GraphNEL object of a KEGG pathway}

\item{Name}{the pathway name}

\item{iter}{the number of iterations for causal disturbance}

\item{alpha}{the dampening factor for Source/Sink Centrality}

\item{beta}{the relative Source vs Sink factor}

\item{statEval}{Choose 1 for product-based, 0 for summation-based}

\item{pathID}{KEGG ID of an input pathway}

\item{deKIDs}{list of differentially expressed genes}

\item{allKIDs}{list of all genes with their associated KEGG IDs}

\item{keggRef}{list of reference genes from KEGG}
}
\description{
Calculates some information regarding the pathways
}
